+++ src/main/java/com/beside/ten011/user/controller/OauthController.java
@@ -1,5 +1,7 @@
-package com.beside.ten011.user;
+package com.beside.ten011.user.controller;
 
+import com.beside.ten011.user.controller.dto.LoginResponse;
+import com.beside.ten011.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
