+++ src/main/java/com/beside/ten011/user/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.beside.ten011.user;
+package com.beside.ten011.user.repository;
 
+import com.beside.ten011.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
