+++ src/main/java/com/beside/ten011/user/service/dto/KakaoProfile.java
@@ -1,4 +1,4 @@
-package com.beside.ten011.user;
+package com.beside.ten011.user.service.dto;
 
 import lombok.Getter;
 import lombok.ToString;
